requirements.txt
load_dotenv
langchain
openai
pinecone-client
tiktoken
unstructured[local-inference]
pypdf